+++ sqllogictest/src/runner.rs
@@ -849,7 +849,9 @@ impl<D: AsyncDB, M: MakeConnection<Conn = D>> Runner<D, M> {
                 Record::Statement {
                     sql, expected, loc, ..
                 },
-                RecordOutput::Query { error: None, .. },
+                RecordOutput::Query {
+                    error: None, rows, ..
+                },
             ) => {
                 if let StatementExpect::Error(_) = expected {
                     return Err(TestErrorKind::Ok {
@@ -857,6 +859,16 @@ impl<D: AsyncDB, M: MakeConnection<Conn = D>> Runner<D, M> {
                         kind: RecordKind::Query,
                     }
                     .at(loc));
+                }
+                if let StatementExpect::Count(expected_count) = expected {
+                    if expected_count != rows.len() as u64 {
+                        return Err(TestErrorKind::StatementResultMismatch {
+                            sql,
+                            expected: expected_count,
+                            actual: format!("returned {} rows", rows.len()),
+                        }
+                        .at(loc));
+                    }
                 }
             }
             (
@@ -1383,9 +1395,11 @@ pub fn update_record_with_output<T: ColumnType>(
                 loc,
                 conditions,
                 connection,
-                expected: expected @ (StatementExpect::Ok | StatementExpect::Count(_)),
+                expected: mut expected @ (StatementExpect::Ok | StatementExpect::Count(_)),
             },
-            RecordOutput::Query { error: None, .. },
+            RecordOutput::Query {
+                error: None, rows, ..
+            },
         ) => {
             // statement ok
             // SELECT ...
@@ -1393,6 +1407,10 @@ pub fn update_record_with_output<T: ColumnType>(
             // This case can be used when we want to only ensure the query succeeds,
             // but don't care about the output.
             // DuckDB has a few of these.
+
+            if let StatementExpect::Count(expected_count) = &mut expected {
+                *expected_count = rows.len() as u64;
+            }
 
             Some(Record::Statement {
                 sql,